Move syntax into test package object
